py/pyproject.toml
[project]
name = "braintrust-evals-py"
version = "1.0.0"
description = "Python tutorial for AI and LLM evaluations using Braintrust"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
keywords = ["braintrust", "ai", "llm", "evaluation", "python"]
dependencies = [
    "braintrust>=0.0.205",
    "autoevals>=0.0.129",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.ruff]
line-length = 88
target-version = "py38" 
